• Okay so I have a voxelized bisou scene! :D

+++ Tests/TestAreaLight/AreaLightForm.cs
@@ -1674,7 +1674,7 @@ renderProg PostFX/Debug/WardBRDFAlbedo {
 
 				ByteBuffer	VerticesBuffer = VertexPt4.FromArray( Vertices );
 
-				m_Prim_Quad = new Primitive( m_Device, Vertices.Length, VerticesBuffer, null, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.Pt4 );
+				m_Prim_Quad = new Primitive( m_Device, (uint) Vertices.Length, VerticesBuffer, null, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.Pt4 );
 			}
 
 			{
@@ -1686,7 +1686,7 @@ renderProg PostFX/Debug/WardBRDFAlbedo {
 
 				ByteBuffer	VerticesBuffer = VertexP3N3G3B3T2.FromArray( Vertices );
 
-				m_Prim_Rectangle = new Primitive( m_Device, Vertices.Length, VerticesBuffer, null, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.P3N3G3B3T2 );
+				m_Prim_Rectangle = new Primitive( m_Device, (uint) Vertices.Length, VerticesBuffer, null, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.P3N3G3B3T2 );
 			}
 
 			{	// Build the sphere
@@ -1731,7 +1731,7 @@ renderProg PostFX/Debug/WardBRDFAlbedo {
 					}
 				}
 
-				m_Prim_Sphere = new Primitive( m_Device, Vertices.Length, VerticesBuffer, Indices, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.P3N3G3B3T2 );
+				m_Prim_Sphere = new Primitive( m_Device, (uint) Vertices.Length, VerticesBuffer, Indices, Primitive.TOPOLOGY.TRIANGLE_STRIP, VERTEX_FORMAT.P3N3G3B3T2 );
 			}
 
 			{	// Build the cube
@@ -1800,7 +1800,7 @@ renderProg PostFX/Debug/WardBRDFAlbedo {
 
 				ByteBuffer	VerticesBuffer = VertexP3N3G3B3T2.FromArray( Vertices );
 
-				m_Prim_Cube = new Primitive( m_Device, Vertices.Length, VerticesBuffer, Indices, Primitive.TOPOLOGY.TRIANGLE_LIST, VERTEX_FORMAT.P3N3G3B3T2 );
+				m_Prim_Cube = new Primitive( m_Device, (uint) Vertices.Length, VerticesBuffer, Indices, Primitive.TOPOLOGY.TRIANGLE_LIST, VERTEX_FORMAT.P3N3G3B3T2 );
 			}
 		}
